Initial code for the redone githubImport.
Lose ability to import all scripts in a repo at once. The user could middle click the import button to remain on the import page. Import is currently a POST request, but we could change it to a GET request and use an <a target="_blank">.
Add ability to import javascript libaries.
Paginate repos.
We are NOT currently paginating a repo's blob tree.
Repo page does filter out files that are not JS.
Indicate when a file is too large to import.
List a github users repos on one page, and list the repo tree on another.