Misc fixes and style guidelines for PR #129.
Avoid using functions with multiple statements in one line during the async -> render code.
Fix error when visiting the new topic page when not logged in (missing `res.`). Thanks Martii.
Check if the user object is a Mongoose document or a plain object (without the .save() function) before updating the user's role when submitting a script.
Add `by ScriptAuthor` after the script version in a script list.
Use the same bootstrap-markdown.css for Bootstrap 3 on newDiscussionPage that scruotNewIssuePage was using.
Make a new include for a comment with an editor to use on newDiscussionPage and scruotNewIssuePage. Update with changes made to include/comment.html.
Add indicator that importing a script from github will open a new tab.
http://sourceforge.net -> https://sf.net
Make the difference between an unvoted up/down arrow and it's disabled version more distinct.