Fix CIDR
Currently GH uses the [https://api.github.com/meta](https://api.github.com/meta)

``` json
{

    "verifiable_password_authentication": true,
    "hooks": [
        "192.30.252.0/22"
    ],
    "git": [
        "192.30.252.0/22"
    ]

}
```

We currently only allow 256 hosts when the `/22` means allow 1024 hosts/broadcasts ranging from 192.30.252.0 to 192.30.255.255 from them. While we may not need this many just yet there is no guarantee that the first 256 are not in use and we might get a WebHook notice or notice from the other 75% of the missed IPs.