Make userscript search the new default for Chrome's "Tab to search"
When auto-adding a OpenSearch plugin on the first visit of a website, Chrome seems to take the first OpenSearch plugin in the HTML header. This commit changes the order of the OpenSearch plugins in the HTML header, so that userscript search is the first one. Will probably only work for first time visitors of the site.