Still focus on IPv4 but diff dep (#1408)
> IPv4-mapped IPv6 addresses are automatically converted back to IPv4 addresses

This appear to do exactly what we are looking for without the extra baggage. A little newer to the mix and not semver major yet... but seems stabler. Thanks to and welcome "the new kid on the block".

Threw a couple of real IPv6 addresses into the mix and it returned appropriate values:

``` js
ipRangeCheck(
  'FE80::0202:B3FF:FE1E:8329',
  ['192.30.252.0/22', '185.199.108.0/22', '140.82.112.0/20']
) // Should return `false` and does

ipRangeCheck(
  'FE80:0000:0000:0000:0202:B3FF:FE1E:8329',
  ['192.30.252.0/22', '185.199.108.0/22', '140.82.112.0/20']
) // Should return `false` and does

ipRangeCheck(
  '::ffff:185.199.111.0',
  ['192.30.252.0/22', '185.199.108.0/22', '140.82.112.0/20']
) // Should return `true` and does

ipRangeCheck(
  '::ffff:185.199.112.0',
  ['192.30.252.0/22', '185.199.108.0/22', '140.82.112.0/20']
) // Should return `false` and does

```

If we absolutely must... we can go back to re's... but I'd rather not at this time.

Ref(s)
* http://formvalidation.io/validators/ip/ (validator)
* https://www.ipaddressguide.com/cidr (expander/compressor IPv4)

Auto-merge